+++ ledao-system/src/main/java/com/ledao/system/service/ISysJudicialService.java
@@ -5,46 +5,46 @@ import java.util.List;
 import com.ledao.system.domain.SysJudicial;
 
 /**
- * 司法Service接口
+ * 司法拍卖项目Service接口
  *
- * @author lxz
- * @date 2020-06-09
+ * @author ledao
+ * @date 2020-09-14
  */
 public interface ISysJudicialService {
     /**
-     * 查询司法
+     * 查询司法拍卖项目
      *
-     * @param id 司法ID
-     * @return 司法
+     * @param id 司法拍卖项目ID
+     * @return 司法拍卖项目
      */
-    public SysJudicial selectSysJudicialById(Long id);
+    public SysJudicial selectSysJudicialById(Integer id);
 
     /**
-     * 查询司法列表
+     * 查询司法拍卖项目列表
      *
-     * @param sysJudicial 司法
-     * @return 司法集合
+     * @param sysJudicial 司法拍卖项目
+     * @return 司法拍卖项目集合
      */
     public List<SysJudicial> selectSysJudicialList(SysJudicial sysJudicial);
 
     /**
-     * 新增司法
+     * 新增司法拍卖项目
      *
-     * @param sysJudicial 司法
+     * @param sysJudicial 司法拍卖项目
      * @return 结果
      */
     public int insertSysJudicial(SysJudicial sysJudicial);
 
     /**
-     * 修改司法
+     * 修改司法拍卖项目
      *
-     * @param sysJudicial 司法
+     * @param sysJudicial 司法拍卖项目
      * @return 结果
      */
     public int updateSysJudicial(SysJudicial sysJudicial);
 
     /**
-     * 批量删除司法
+     * 批量删除司法拍卖项目
      *
      * @param ids 需要删除的数据ID
      * @return 结果
@@ -52,10 +52,10 @@ public interface ISysJudicialService {
     public int deleteSysJudicialByIds(String ids);
 
     /**
-     * 删除司法信息
+     * 删除司法拍卖项目信息
      *
-     * @param id 司法ID
+     * @param id 司法拍卖项目ID
      * @return 结果
      */
-    public int deleteSysJudicialById(Long id);
+    public int deleteSysJudicialById(Integer id);
 }
